rol-premium

+++ src/routes/products.router.js
@@ -1,8 +1,10 @@
 import { Router } from "express";
 //import ProductManagerFS from "../dao/ProductManagerFS.js";
-import   ProductController  from "../controllers/ProductController.js";
+import ProductController from "../controllers/ProductController.js";
 import { uploader } from "../utils/multerUtil.js";
 import productModel from "../models/productModel.js";
+import { verifyToken } from "../middlewares/verifyToken.js";
+import { handlePolices } from "../utils/authUtil.js";
 
 const router = Router();
 
@@ -12,12 +14,12 @@ const productManager = new ProductController();
 
 router.get("/", async (req, res) => {
     try {
-      let {limit = 10, page = 1, query = {}, sort = null } = req.query;
-      const result = await productManager.paginate(limit, page, query, sort);
-      res.send({
-        status: "success",
-        payload: result
-      });
+        let { limit = 10, page = 1, query = {}, sort = null } = req.query;
+        const result = await productManager.paginate(limit, page, query, sort);
+        res.send({
+            status: "success",
+            payload: result
+        });
     } catch (error) {
         console.error("Error al traer los productos:", error.message);
         res.status(500).send({ error: "Error al traer los productos" });
@@ -40,7 +42,7 @@ router.get("/:pid", async (req, res) => {
 });
 
 
-router.post('/', uploader.array('thumbnails', 3), async (req, res) => {
+router.post('/', verifyToken, handlePolices(['admin', 'premium']), uploader.array('thumbnails', 3), async (req, res) => {
 
     if (req.files) {
         req.body.thumbnails = [];
@@ -48,7 +50,8 @@ router.post('/', uploader.array('thumbnails', 3), async (req, res) => {
             req.body.thumbnails.push(file.filename);
         });
     }
-
+    //agrego el atributo owner 3erintegra
+    req.body.owner = req.user.role === 'premium' ? req.user.email : 'admin';
     try {
         const result = await productManager.createProduct(req.body);
         res.send({
@@ -63,7 +66,7 @@ router.post('/', uploader.array('thumbnails', 3), async (req, res) => {
     }
 });
 
-router.put('/:pid', uploader.array('thumbnails', 3), async (req, res) => {
+router.put('/:pid', verifyToken, handlePolices(['admin', 'premium']), uploader.array('thumbnails', 3), async (req, res) => {
 
     if (req.files) {
         req.body.thumbnails = [];
@@ -86,21 +89,31 @@ router.put('/:pid', uploader.array('thumbnails', 3), async (req, res) => {
     }
 });
 
-router.delete('/:pid', async (req, res) => {
+router.delete('/:pid', verifyToken, handlePolices(['admin', 'premium']), async (req, res) => {
 
     try {
-        const result = await productManager.deleteProduct(req.params.pid);
-        res.send({
-            status: 'success',
-            payload: result
-        });
-    } catch (error) {
-        res.status(400).send({
-            status: 'error',
-            message: error.message
-        });
-    }
+        //3era integradora
+        const pid = req.params.pid;
+        const email = req.user.email;
+        let proceeDelete = true;
 
-});
+        if (req.user.role === 'premium') proceeDelete = await check(pid, email);
+
+        if (proceeDelete) {
+            const result = await productManager.deleteProduct(req.params.pid);
+            res.send({
+                status: 'success',
+                payload: result
+            });
+            }
+        } catch (error) {
+            res.status(400).send({
+                status: 'error',
+                message: error.message
+            });
+
+        }
+        
+    });
 
 export default router;
